requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Technical analysis
pandas-ta>=0.3.14b0
ta-lib>=0.4.28

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Database
sqlite3-python>=1.0.0

# CLI
click>=8.1.0
tabulate>=0.9.0

# Configuration
pyyaml>=6.0

# Logging
loguru>=0.7.0

# Task scheduling
apscheduler>=3.10.0

# Data validation
cerberus>=1.3.5

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Web API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Optional: Dashboard
streamlit>=1.28.0
plotly>=5.17.0
